+++ src/main/java/com/hrtx/web/pojo/PinganRefund.java
@@ -11,7 +11,7 @@ public class PinganRefund extends BasePojo implements java.io.Serializable {
 	private static final long serialVersionUID = 1L;
 	@Id
 	@JsonSerialize(using = ToStringSerializer.class)
-	private Long id;
+	private Integer id;
     private String open_id;
     private String timestamp;
     private String sign;
@@ -31,7 +31,7 @@ public class PinganRefund extends BasePojo implements java.io.Serializable {
     public PinganRefund() {
 	}
 
-    public PinganRefund(Long id, String open_id, String timestamp, String sign, String sign_type, String out_no, String refund_out_no, String refund_ord_name, String refund_amount, String trade_account, String trade_no, String trade_result, String tml_token, String remark, String shop_pass) {
+    public PinganRefund(Integer id, String open_id, String timestamp, String sign, String sign_type, String out_no, String refund_out_no, String refund_ord_name, String refund_amount, String trade_account, String trade_no, String trade_result, String tml_token, String remark, String shop_pass) {
         this.id = id;
         this.open_id = open_id;
         this.timestamp = timestamp;
@@ -49,11 +49,11 @@ public class PinganRefund extends BasePojo implements java.io.Serializable {
         this.shop_pass = shop_pass;
     }
 
-    public Long getId() {
+    public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
